ORM->Native sql + refactoring
