+++ src/model/base-dto.ts
@@ -1,3 +1,5 @@
-export abstract class BaseDto {
+import { AbstractDomainEntity } from "nodets-ms-core/lib/models";
+
+export abstract class BaseDto extends AbstractDomainEntity {
 
 }
